Moved my generator into namespace bertof, fixed interface non public
