+++ CMakeLists.txt
@@ -11,7 +11,7 @@ find_package(Boost 1.63.0 COMPONENTS filesystem date_time)
 if (Boost_FOUND)
     include_directories(${Boost_INCLUDE_DIRS})
 
-    set(SOURCE_FILES main.cpp Pin/Pin.h Generator/Generator.h Exceptions/NotImplementedException.h Generator/Generator.cpp Logger/ScreenLogger/ScreenLogger.h Logger/ScreenLogger/ScreenLogger.cpp Exceptions/InvalidInputException.h Generator/GeneratorInerface.h Logger/Logger.cpp Logger/Logger.h Logger/TextLogger/TextLogger.cpp Logger/TextLogger/TextLogger.h)
+    set(SOURCE_FILES main.cpp Pin/Pin.h Generator/bertofGenerator/Generator.h Exceptions/NotImplementedException.h Generator/bertofGenerator/Generator.cpp Logger/ScreenLogger/ScreenLogger.h Logger/ScreenLogger/ScreenLogger.cpp Exceptions/InvalidInputException.h Generator/GeneratorInerface.h Logger/Logger.cpp Logger/Logger.h Logger/TextLogger/TextLogger.cpp Logger/TextLogger/TextLogger.h)
     add_executable(WPS_pin_generator ${SOURCE_FILES})
